src/requirements.txt
flask
Flask
flask-cors
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv